% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erf.R
\name{erf}
\alias{erf}
\title{erf}
\usage{
erf(type, x, dist, ...)
}
\arguments{
\item{type}{choose the type of evaluation}

\item{x}{Value to calculate the evaluation.}

\item{dist}{Distribution that you want to use.}

\item{...}{Specific arguments of your model.}
}
\description{
erf
}
